Client - Separate Type definitions Into Different Files
Create new files for CPU, memory, Disk, and Wifi
