+++ client/src/types/common.d.ts
@@ -1,4 +1,1 @@
-export interface IResponse<T> {
-  Results: T[]
-  Total: number
-}
+
